move format functions from segmentation to stack
